Pyehtim

Documentation for Pyehtim.

Pyehtim.get_arraytable

Pyehtim.get_datatable

Pyehtim.get_fr_angles

Pyehtim.get_scantable

Pyehtim.load_uvfits_and_array

Pyehtim.get_arraytable — Method

getarraytable(obs)

Construct the array table for a given eht-imaging obsdata object.

Pyehtim.get_datatable — Method

get_datatable(obs)

Extracts the eht-imaging obs.data into a Julia table. Note this copies the data. For a non-copying version just access data the usual way

Pyehtim.get_fr_angles — Method

get_fr_angles(obs)

Returns the feed rotation angles for a given observation as a set of tuples. The first tuple contains the elevation correction for site 1 and 2. The second tuple returns the parallactic angle correction for site 1 and 2

Pyehtim.get_scantable — Method

get_scantable(obs)

Constructs the scan table of a given observation.

Pyehtim.load_uvfits_and_array — Function

load_uvfits_and_array(uvfile, arrayfile=nothing; kwargs...)

Load a uvfits file with eht-imaging and returns a eht-imaging Obsdata object. You can optionally pass an array file as well that will load additional information such at the telescopes field rotation information with the arrayfile. This is expected to be an eht-imaging produced array or antenna file.